FIX:refactored post routes code from normal routes to controller
